Speed up large io calls with Reader Tools
BL-3542 was largely caused by trying to do some 1.4mb write on the UI thread. Possibly this is becuase the client is running on the UI thread too, and the size of the payload was large, so the writer couldn't write because it was waiting for the client to consume, but the client couldn't because the thread was blocked waiting to write some more...

Anyhow so I made it possible when registering a handler to say that it doesn't need to be run on the UI thread.

I then split all the READER api calls into either "io" or "ui", and made separate registrations for the two groups.