"Object is currently in use elsewhere" making thumbnails
From the stacktrace, I saw that the contention was in the Clone(). This image is cached by the thumbnailer, so it appears that multiple server threads were asking for the same image and being cloned at the same time. But the Clone() itself appeared to be a previous attempt at preventing this kind of error, probably in the Save(). A better solution, it seems to me, is to lock() on the image  while we save it.