Fix epub tests
Added languagedisplay.css to tests. Had to hold my nose while doing this, it feels like a hack. Every unit test should not have to be updted to whatever the current stylesheet system is; but the system of detecting which fields are visible actually *goes around* the stylesheets of the actual page, so that the only stylesheets it sees are what have been added to the unit test. Very frustrating, but not worth more time now.