Optimize page switching (#1714)
* Avoid loading page chooser into main page

* Cache final file path for each url

When we've once determined that a particular file should be returned in response to a particular URL, remember that and return it at once in future

* Don't delay loading main page

Only two things really need the toolbox to finish loading before the content page; this tweaks them to allow both to load from the beginning.

* Update only page iframe on change page within book

* Allow browser to cache files unlikely to change

* Remove the new local server caching

Obsolete now we allow the browser to cache.
The minor enhancement of using Response.Close(data) remains.

* Cleanup following review

* More post-review improvements