Extract xliff from pug and markdown files
Add the generated xliff files to source control for crowd-in to find.
As long as the pug and markdown files don't change, and the HtmlXliff
program doesn't change, generating these files afresh shouldn't cause
any differences.

This is the first step of a projected 3-step process.  The second step is
to extract translated xliff files from the translated HTML related file.
The third step is to actually change the program to remove the translated
pug and markdown and create the translated HTML files at build time and
add them to the installer and package.  The program behavior (and maybe
the content of the installer) won't change until the third step.