Cleanup Android Publishing Classes
This moves all the logic of each method out to its own class(es), leaving the API to simply run things and manage the state machine, and hide the whole http request/ web socket layer from the classes that do the publishing. With the new UI, the state system is a lot simpler, so that allowed some simplication too.