Fix Usb Sending (that I broke) BL-4927
I had broken this when I changed the state machine so that "connecting" and "sending" are not two different states. Once we connect, we now immediatly send.

While trying to understand how it was supposed to work, I did some simplification here and there. 

The extra interface and dummy class were to make things compile on Linux, but we think that can be handled via #ifdef's. 

As far as I can tell, real "events", which are a bit harder to work with, weren't needed.