Fix progressDiv null error BL-4929
The real error turned out to be a simple change of id. But before I figured that out, I had made the code a bit more robust. Since there I found an instance of this error in the youtrack history, unsolved, it's possible that this will be useful, so I left that in there.