Simplify the WebSocket management system
With this commit, users of WebSocketManager have to supply a socket name. The ability to send a message that closes down a named socket is retained, but is not currently used as it did not appear to have any measurable effect, for good or ill. 

On the c# side, we now close sockets instead of just removing them from a list. 

For ease in debugging, on both sides, socket names are now available, by hijacking the "subprotocol" field.