Fix BL-4901
The problem was that the c# reply to the POST did not contain any text (not even an "OK"). I don't know why it should, actually. But without that, the JS POST eventually throws the error we were seeing in BL-4901.

In order to prevent this error in the future, I have renamed 2 of the methods so that it will be clear which one to use:

request.SucceededDoNotNavigate() --> request.ExternalLinkSucceeded()
request.Succeeded() --> request.PostSucceeded()