Use MarkdownDeep like libpalaso instead of Markdig (BL-5005)
This prevents a crash when showing the Release Notes dialog.  (We still
need the reference to Markdig for the benefit of HtmlXliff.  This will
go away as more progress is made on the new localization process.)

For Linux, obtain NDesk.DBus.dll from libpalaso as well as MarkdownDeep
to allow USB accessing code in libpalaso to work in Bloom.