Bl-6273 experiment: AngleSharp
I got AngleSharp working, but the experiment just convinced me that we this whole area of code needs an overhaul, because it is possible to have mutations of the that trigger errors before we ever get to ValidateBook, which do not then lead to the use of the .bak file. See https://issues.bloomlibrary.org/youtrack/issue/BL-6273

So I'm just putting my experiment here for when we cycle back to this issue.