wip respect data-div when changing to a different page layout
Putting this away for now. It was a spike that proved this isn't super cheap to pull off, but is conceptually sound. committing this work so we can return to it.

The idea is to allow a template page to contain data-book elements. This would allow, for example, the first page of a story to be modified to have the story's title and credits at the top. This is desired when compiling lots of stories together into a single "reader".

The code here does allow that, but the values from the template page's fields with a data-div (e.g. bookTitle, bookSmallCredits) are being read from the new page. Bloom them overwrites  those variables in the destination book, so the book takes on the name of the template!

I've added a sample template to this commit, but it's just a hacked html, no pug or anything.