Update TS to use arrow functions, part 3 (BL-6401)
This is the final set of files, including the change to tslint.json to
detect places where arrow functions should be used.  Note that tslint
ignores function declarations whose body already uses a this variable
different than the this in the outer scope.  These places may warrant a
second round of fixes, but will require much more thought and care.