Update PNG-RISE branding and Release Notes
Note that the landscape one has the EFL logo, and the portrait doesn't. The EFL unit apparently only uses the landscape.
Also, note that they only do electronic versions and this only fits when used with the "Device" xmatter. The layout it what they want, I only cleaned up the logos and alignment a little bit.
The SVGs were exported from figma without the outline text since the are Andika and this is built in, thus saving a lot of space.