Convert Callout Tool from React Class Component to React Hooks
Early Prototype of Callout Tool using React hooks

Turning on/off bubble editing when the toolbox is opened/closed works. Can switch between bubbles on canvas and it'll be updated in toolbox. You can update the options in toolbox and it'll update the canvas, but some wonky bubble-switching behavior if modifying not the 1st bubble selected.