Fix minimal coverage of book in Book.AllLanguages (BL-7800)
This doesn't really address the issue in BL-7800 but is a needed fix.
Without, if the first div.bloom-translationGroup encountered had all
languages in it, it would be the only one checked in the whole book.  The
previous code only saved the parent div for the first div.bloom-editable
that had content for a given language.