Various improvements to publish to Bloom Reader and related (BL-7892)
Adds a Refresh button
Removes automatic refresh when toggling a language check box
Various tweaks to the layout of the check boxes
For BloomReader, allow preview to be generated even if something goes wrong with navigating to the page we need for figuring out invisible stuff to delete (just don't delete it)
Fix up the way we release simulated page data to prevent race conditions releasing it too soon.
Fix a race condition that could prevent navigating to the is-visible page at all.