Handle customized language names more consistently (BL-7832)
A customized name will always be used, in the UI and published material.
A non-customized name will depend (in theory) on the context, the UI
language, and the national language for the collection.  In practice at
the moment, we have only possibly custom names, English names, and the
name of the language in itself (autonym).