Use language.Name for the (up to) three chosen languages  (BL-8302)
This gives the user control over what is seen on the screen and what
is published in the book without any surprising changes in some places.
This breaks the fix for BL-7382 until the libpalaso language chooser
dialog is changed to offer autonyms instead of (or in addition to?)
English names.