Properly set version number during package build
The package build, which runs inside of a chroot environment,
didn't have access to the environment variable that we tried to
set before we started the package build. The result was that the
patch version was always 0.

This change depends on a change in ci-builder-scripts that stores
the build number in `build_number.env`.