Fix to improve handling of hard crashes (BL_8619)
This fix increases the chance of getting a stack trace and error report
for access violation crashes which otherwise cause the program to simply
disappear.  It's still not foolproof since the action of bringing up the
text editor with the stack trace can still cause another access violation
which is unhandled.  But it improves the odds considerably.