Tweak fix for cropping only second time bug (BL-8638)
Looking at the code again, this seems like an obvious fix.  The current
implementation of the the dialog makes the current code work okay, but
what if the implementation changes to always instantiate a new ImageInfo
when the image changes in any way, not just a new Image in the ImageInfo?