Talking Book: Fix bug where backspace unnecessarily prevented (BL-8828)
If you have Talking Book open in Record by Sentence mode, then position the cursor at the start of the 2nd sentence of the 1st paragraph, then press backspace...
Previously it would prevent the backspace. (Because it has offset 0 and is inside the 1st paragraph).
Now we do a more sophisticated check of whether it's at the beginning or not.