Talking Book: Fix bug for multiple text boxes (BL-8797)
When you Listen to the whole page, they are played in sequence. After the 1st text box finishes, the 2nd text box starts playing its audio, which updates the audioPlayStartTime (good). The 1st audio is emptied and triggers an onemptied event. When onemptied is handled, it sets audioPlayStartTime to null (but audioPlayStartTime had already been updated to the correct value!), so you end up in a bad state.