Reenable a number of tests on Linux (20201020)
The reasons for disabling them appear to no longer apply.  They all run
okay on the local machine.
I added a "SkipOnTeamCityLinux" category for the one test that runs okay
locally but would produce an error from a warning message on TeamCity.
It is possible that one or more of the reactivated tests may fall into
that category.