Have Android depend on user settings for blind and sign language (BL-8995)
Creating bloomd books through harvester were setting signLanguage as a
feature whenever a video exists in the book.  This was then getting set
in the parse table, and showing in next.blorg.  Blind and sign language
should be set only by the user when uploading.  (Talking book is still
being set automatically as a feature: it will be controlled by the user
allowing narration audio to be uploaded.)