Fix arithmetic templates overflowing (BL-9102).
If you tab into an Equation-style editable, then type into it, the selection is before the first child (aka the <p> tag). However, we would like the selection to be at the start of the <p> tag, not before.  For bloom-editables, this is normally taken care of by CK Editor, but .Equation-style editables were previously not handled by CK Editor.