BL-9912 saving tool state too late
If a save to tool state comes in after we're already back on the collection tab, it's not going to get saved anyways. And it seem to not matter, since important tool state gets saved as it happens (e.g. reader level). A better fix JohnT and I decided is way too expensive.