Nag user about concurrent updates only when problematic (BL-10227)
With React asynchronously asking for HTML which can trigger a preview
update, creating a new book from an existing one often updates the
preview DOM and the real DOM simultaneously.  This seems harmless to
me as the code is operating on different input data.  The master branch
has also added more calls to BringBookUpToDate, such as from SelectBook.