FFmpeg configure refactoring: Remove redundant disable library options
disable-autodetect covers most of the manually specified disable options. Actually, not having those extraneous disable options REDUCES the file size (only by a tiny amount), counterintuitively. The only purpose of those disable commands is to reduce the file size, so if it's not reducing it, I think it's perfectly fine to delete them. My hypothesis for this counterintuitive effect is that since the ffmpeg program does print out the options it was configured with, reducing the text length of the configuration settings has a tiny effect on the file size.
Note: --disable-libxcb-shape is not technically covered by disable-autodetect, but libxcb-shape is disabled by default, so it can be removed.