Restore NotifyUserOfProblem allows "Report" by default (BL-11193)
In a recent commit, the code was changed so that when the exception was null, the "Report" button was not presented by default. (This was more in line with how libpalaso's ErrorReport works). But this was a change compared to how Bloom's HtmlErrorReporter was supposed to work.
So, undoing that change in behavior.