Rename files to trigger branding file updates (BL-11543)
When merely fixing capitalization, the branding files weren't automatically updated. This is probably because Windows (being case-insensitive) doesn't register any difference between the old and new filenames, which differed only in case.
A workaround to this is to just append "-logo" to the filenames, making it obvious to Windows that it should update the branding files.