Consolidate code that handles Shift-Enter linebreaks (BL-11717)
Based on Steve's commits 73404d1c6e768556c10ef5446d1bd20fa1b198dc and 847bc6caa3b9bc55b84351289db9042aa6b555e6. However, that version had issues in one of the unit tests because XmlDocument and XElement handled newlines differently.  I re-did it a couple weeks ago to utilize the Book.cs version as the primary (instead of BookData as the primary in Steve's), which addresses the unit test problem.
This commit merges my prior work with the less-risky version of the change made to the 5.4 branch.