Switch to server finding images by exact paths (BL-9989)
After the change, images that are part of the contents of a book are expected to be in the root folder of the book; images (like UI icons) that are part of Bloom's implementation must be at the specified path relative to output/browser, with a handful of exceptions.
As well as improving performance, this prevents finding unintended files in obscure locations when the file we want is not where we expect.