More post-review cleanup
We use MUI Tooltip to allow checkbox and tickedBox tooltips to be any react node, which allowed moving the link that shows the current sign language (or to set it if missing) into the appropriate tooltips.
Got rid of the spurious new publish setting for visually impaired and made both the old and new control set it for the current language 1, not for a random one or all of them.
Pulled Blorg SVGs into own files.
Many other tweaks.