stop using pretty-quick
pretty-quick is unmaintained and was giving me errors with latest prettier. we already have lint-staged, and it can run prettier. However, it has to write, and the comments make it clear that Jeffrery intentionally set lint-staged to "no-stash" which means we can't write. I just decided to give up on prettifying css/less for now; nowadays, vscode should be set up to do that for everyone.