Add data-default-languages-legacy (BL-12975)
Adds this attribute for use on fields (currently just .Title-On-Cover-style) that previously used data-default-languages to control which languages should be visible (by adding the class bloom-visibility-code-on). The value in the new attribute, if present, will be used when the book is in legacy mode. Otherwise, the bloom-visibility-code-on class will be removed.
Also removes all reference to bloom-visibility-user-on/off, an idea we introduced with bloom-visibility-code-on but never used.