Harvester uses legacy theme (or fails if that is not allowed) (BL-13174)
The essence of this change is the code in AppearanceSettings that always uses legacy theme if running in harvester mode, the code in BookStorage that throws a special exception if running in harvester mode and not allowed to use legacy, and the code in CreateArtifactsCommand which catches that exception and reports a new error code. I also created a unit test and had to change a few things to let it work; and discovered that one of the arguments to the BookStorage Factory was unused, and removed it from there and various related methods.