Fix code and tests of getRecordableDivs
Originally we had two versions of the logic to get elements for which talking book should offer to record audio, one in getRecordableDivs/getAudioElements and one in setCurrentAudioElementToFirstAudioElementAsync. For Bloom games, I added checks for bloom-visibility-code-on and not in a target and a more thorough check for any parent with display:none, and refactored to have the logic in one place. I missed that unique to the setCurrentAudioElementToFirstAudioElementAsync version was a check for being inside a translation group.
Once it was all consistent a number of tests had to be enhanced to have translation groups and visibility code set.
Finally, had to do cleanup to make our tslint commit hook happy.