Make data-sound work in Bloom Player for non-activity pages (BL-13878)
The change to dragActivityRuntime is not needed in BD, but the files are supposed to stay in sync between the two repos, at least until the new games code stabilizes.
Putting this in draft, because I want it reviewed, but to make this work fully, we need to update BD to use the new BP that will emerge when that PR is merged.