Some fixes and possible fixes
John had a couple of problems. One was a failure to detect that a mouse click on the control bar should not remove it, because the previous check depended on it being a child of the control frame. Another somehow left things in a state where a bubble was being dragged with no button pressed. I decided to make that impossible. In the process I found a whole lot of obsolete resize code to remove.