Manage focus and activeElement for prompt dialog
Prevents problems with leaving the controls on a deleted node; a click to focus the prompt dialog causing the control frame and focus to shift (instead, we remove the frame when opening the dialog).
I also improved error handling for the mess that gets created if we accidentally import editablePage.ts into some other bundle.