WIP Improve resizing of overlays when container size changes (BL-13965)
The basic algorithm for resizing is working well, I think.
New tests are still sparse.
I think rather than treating the background image specially, we want to transform it into an overlay (and add a new placeholder) when we go into this mode. Then we get all the overlay controls for image overlays for free, and better backwards compatibility. We may not even need the new class.