Adjust better when image container size changes (BL-13965)
This converts the background image of an image container into an overlay, allowing it to be cropped.
The main image is left as a placeholder, so earlier Bloom versions won't find things too surprising.
We still have a way to identify one overlay as a background, with some special behaviors.