Force correct interface to be used, finally
I put the relevant code in a small console app to figure out the socket stuff. When it worked I ported it back into the BloomDesktop code. The most recent thing to overcome was developing a way to generate the broadcast address to use with a given interface, since all-FF's gave me a permissions violation in the console app.
Next steps: rather than hardcoding localIp and subnetMask, use the IP returned by GetIpAddressOfNetworkIface(). Then search among the interfaces to find the corresponding netmask.