BL-14363 Show Expiration Notices
There are also a number of eslint-related things I had to do to get it to compile. Still a mystery why this doesn't happen to other people.

There are a number of things that aren't as clean as I could want, but a) this is temporary and b) we are slipping this into a release version, so I don't want to go about refactoring.