Fix bug in broadcast address computation
Broadcast address is incorrectly calculated for some instances when the subnet mask has less than 24 '1's. Replace with a different algorithm that correctly handles all (I think) cases. Not sure it is reasonable to test all possible combinations of IP addresses plus subnet masks, but this algorithm corrects the errors I observed.
This is the code after the first test, which was successful. Still need to remove the debug output and, probably, replace the new throw with a return plus an error to be logged.