Broadcast address computation fix - cleanup
Removed debug output. Fixed the new catch block to return an empty string instead of an error message. Replaced the throw with an error to be logged and an empty string return to be consistent with the rest of this WiFi code base.
Still need to test on the larger JAARS network before updating the PR (#6811) with these changes.