Fix overflow sizing glitches while loading page (BL-14489)
Don't limit bloom-show-en-when-blank to drag-activity-start. It could be useful elsewhere, and we get flicker because some code that depends on it runs before drag-activity-start gets set.
But there are other sources of flicker from checking overflow too soon. And we had two separate sources initiating overflow checks and two different ways of delaying the checks and making them incremental. So I removed the old one that was happening too soon.