Make letter boxes grow wide enough for letters (BL-14541)
BL-14541 grew to be much more complex than this.
The current state of this card is handling width fairly well, but failing to get a reasonable height for မြန်မာ unless line height is 1.5, in which case ordinary roman letters overflow mysteriously. Something is wrong with measuring text. There are also many related tasks in the card which are not yet addressed here.