Put recapitalized language tag back together manually (BL-14597)
Using the SIL.WritingSystem.IetfLanguageTag function could lead to pieces
of the input tag being removed because they are implied by default, but
that can lead to unexpected problems.  Recapitalization works okay because
tags are somewhat case insensitive, but removing subtags is a major change
that makes the tag unrecognizable.