Reorganize UpdateSupport (BL-15145)
Not sure I actually fixed any bugs, even BL-15145, but it was getting hard to reason about all the states and paths in the update code. I cleaned it up so the possible states are explicit, there is basically only one sequence of steps through the upgrade process, and Check For Updates will always report where you are in the process. Added some exception handling that might help.