Removes a lot of files we don't need (BL-15245, BL-15266)
A bunch of 32-bit DLLS some of which may have been needed in older versions of Windows and/or with dependencies we've retired.
A bunch of runtimes for platforms we aren't trying to support.
(Not sure if having the ARM runtimes is helpful. I think we would need a native ARM version of Bloom.exe, at least.)