Last debug commit: comments and debug output
Improve a few comments. Tweak a bit of debug output.
The new advertising loop, in which IP address is discovered before EVERY advert, tests okay. If an advertising session begins with Wi-Fi but midstream changes to Ethernet, WiFiAdvertiser gets it and adjusts. Same thing for the opposite case, Ethernet > Wi-Fi.