Final cleanup in preparation for pull request
Removed all Debug.WriteLine() calls. Reformatted some calls that somehow got automatically formatted (did csharpier do this?) I'm trying to minimize the diffs to be viewed in what will be the second version of this PR.
There is one substantive change: added a dispose call on the UdpClient in Stop(). Visual Studio asked for this since WiFiAdvertiser now inherits from IDisposable (which was a main part of the merge conflict).